bitbake: fetch2/git: allow override of clone args with GITCLONEARGS
Currently there is no way to avoid the "--mirror" argument when doing an
initial clone.  This is problematic in that there are now a significant
number of repos out there for which we want the content, but creating a
full mirror clone isn't practical or efficient.

The two most common examples are:

1) Repos that have "ingested" another repo with no common ancestors - we
see this in "stable-queue" and "linux-rt-devel" repos[1] which started
as repos of queued format-patch output, but then also added the baseline
so the patches could be obtained "pre-applied" for easier automated testing.
Obviously, ingesting all of the linux history grows a repo by gigabytes.

2) Repos that have gone forward in time in a direction we don't care about,
say for example a flag day event, where everything was rewritten from C into
rust in one massive commit, or where a merge of a giant wad of orthogonal
history and/or binary/proprietary content "pollutes" the repo.

In both cases, we can solve the problem by using "--single-branch"
optionally in conjunction with "--branch <goodstuff>" in order to limit
the cloned content to a specific branch or tag.  We will only get
content up to that point, and any commits/ancestors leading up to that
point.  The clone is fully functional and without any of the limitatons
imposed by shallow clones.

So, in case #1 we simply choose the branch we want - raw unapplied
content, or applied and tagged (and perhaps signed) - but we aren't
subjected to downloading both branches.   In case #2 we simply specify
the last known "good" tag before the upstream went off in a direction
that we don't care about.

Note that default behaviour is unchanged - so this is an opt-in feature
which won't impact any existing recipes.  We also leave the door open
for it being URL specific via <name> since we know we have multi repo
SRC_URI in recipes-kernel already.

[1] git://git.kernel.org/pub/scm/linux/kernel/git/rt/linux-rt-devel.git
    git://git.kernel.org/pub/scm/linux/kernel/git/stable/stable-queue.git

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>