bitbake: fetch2/git: allow optional git download name overrride
Single repositories containing multiple independent origin commits are
becoming more common.  For example the rt development repo has patches
as "raw" format-patch output in a patch series, and also for ease of
automated testing, them all pre-applied (ff and non-ff) to the original
baseline (v5.10.8 in this case):

   linux-rt-devel$ git merge-base linux-5.10.y-rt linux-5.10.y-rt-rebase
   7a1519a74f3d0b06598fb95387688cde41e465d8
   linux-rt-devel$ git describe 7a1519a74f3d0b06598fb95387688cde41e465d8
   v5.10.8
   linux-rt-devel$ git merge-base linux-5.10.y-rt linux-5.10.y-rt-patches
   linux-rt-devel$
   linux-rt-devel$ git log --oneline linux-5.10.y-rt | tail -n1
   1da177e4c3f4 Linux-2.6.12-rc2
   linux-rt-devel$ git log --oneline linux-5.10.y-rt-patches | tail -n1
   96f1ac50158c [ANNOUNCE] 4.1.2-rt1

As can be seen, there are two different "epoch" commits and no history
or shared objects exist between the patches branch and the applied
branches (on the linux-stable baseline).

Since we construct the download directory name based on the server name
and the server's path to the repository, we currently can't refactor that
content into two locally independent paths/repositories in the download
dir for efficiency, ease of use and tarball/MIRROR handling.

To resolve this, we allow the recipe to optionally specify the path
component of the download dir name, so that a server with a repo of N
independent branches/histories can be used in N different recipes (or N
different SRC_URI lines) without overlap or name space collision in the
download dir.  Continuing with the above example, we could have:

  RT_REPO = "git://git.kernel.org/pub/scm/linux/kernel/git/rt/linux-rt-devel.git"
  SRC_URI = "${RT_REPO};name=applied;dlname=.preempt-rt.linux-5.10.y \
             ${RT_REPO};name=patches;dlname=.preempt-rt.patches-5.10"

...resulting in the following independent repository dirs:
     downloads/git2/git.kernel.org.preempt-rt.linux-5.10.y
     downloads/git2/git.kernel.org.preempt-rt.patches-5.10

If the new "dlname" parameter is not specified, then the automatic name
generation functions just as it always has done and this change has no
effect on any existing download dirs.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>