bitbake: fetch2/git: ensure static repos have at least one refs/heads
When you clone with "--single-branch --branch foo" one of two things
will happen:

If foo is a branch upstream you will get a local branch foo and HEAD
will contain "ref: refs/heads/foo".  Output from "git branch" will be:
	* foo

If foo is a tag upstream, you won't get any local branch, and HEAD
will contain the SHA of the tag foo.  Output from "git branch" will be:
	* (no branch)

In the latter case, people tend to assume it is some kind of indication
progress was interrupted and an error happened.

While not strictly necessary, in static repos, we point master at the
tag that was fetched if no branches are present, to avoid confusion.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>