kernel: add basic boilerplate for fetch-only recipes
With changes to the fetcher, we now have what we need to download
kernel sources in a way that is forward thinking for sharing and
growth and EOL of components.  This will be achieved by "fetch-only"
recipes that only exist to populate chunks of this "library" of source.

We may wish to reconsider locating 99% of this elsewhere, if use cases
outside the kernel eventually arise.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>