kernel: add a fetch-only recipe for mainline v5.10 source
With v5.10 being the newest baseline currently in use by linux-yocto and
with the download size being 1/2 the size of current linux-yocto itself,
the v5.10 makes a good initial line in the sand for blocking out the
source in a way to optimize sharing.

With this commit present, one can test via "bitbake -c fetch linux-5.10".

Pending recipes who will want this source present will add:
	do_fetch[depends] += "linux-5.10:do_fetch"

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>