kernel: allow splitting mainline v5.10 source download in two
With support for sharing packs, we can now optionally split the larger
mainline kernel content up to v5.10 (approx 1.5G) into two.  This may be
advantageous to people with release media constraints or unreliable
connections.

By setting "INITIAL_KERNEL_SPLIT = 1" in your configuration, you will
get a linux-4.0 clone and then the linux-5.10 clone in your downloads
directory.  From a size perspective, v4.0 is roughly the half way point
between when git use started in 2005, and v5.10 from 2020.

This split is achieved by manually selecting the midpoint tag and cloning
that portion 1st, and then cloning the 2nd 1/2, while using the packs in
the 1st 1/2 as a reference.

Packs deploy both zlib and delta compression - and as with all
compression technologies, splitting the input comes at a cost in terms
of reducing the compression.  So two packs will be bigger than one,
three packs bigger than two, and so on.  Representative numbers are
listed alongside the implementation in the change itself.

The 2nd 1/2 includes a fetch dependency on the 1st 1/2, so that things
will be serialiazed properly.  The dependency also links it into the
chain of events, so no other mentions of v4.0 are required.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>