kernel: allow splitting mainline v5.10 source download in three
With a split in two done, extending to three is fairly trivial.  The
only difference is that the introduced "touch down" point of 4.12 has a
fetch depend/ref itself to the other touch down point of 3.8.  The parent
repo (v5.10) has no need of knowledge of what happens prior to 4.12.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>