kernel: add recipe for linux-master (mainline latest)
The idea for this recipe is to have a name "linux-master" that people
can depend on as a reference for mainline; today and ideally for as long
as we are maintaining it.  While the recipe is there for everybody,
the obvious user we are creating this for is the linux-yocto-dev repo.

At some point in the future, we'll carve off another chunk of post-v5.10
work, mark it static and put it alongside of v5.10 and let this mainline
recipe reference that.  But use of this "linux-master" will be 100%
transparent to the underlying backing store changes.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>