kernel: add stable fetch recipes for v5.4.x, v5.10.x and v5.12.x
One might wonder why there isn't just a stable-5.x to encompass all of
these, but by having them separate, we can retire/EOL them individually
without disruption.

Note that references are a "round up" to the nearest kernel recipe
containing the stable baseline - so stable-5.4 references linux-5.10 as
we didn't specifically call out linux-5.4 as an independent recipe.

Also, stable-5.12 doesn't exist yet, but we know it will, so we insert
a placeholder in for it, so that v5.12-rt can reference it w/o future
change; we just switch the one line in stable-5.12 once it exists.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>