kernel: add preempt-rt fetch recipes for v5.4.x, v5.10.x and 5.12.x
Note that the v5.12 content currently is in the devel repo and will
eventually get moved over to the stable-rt repo.  When that move
happens, we'll need to update the SRC_URI accordingly.  [The v5.12-rt
content itself will remain fast-forward with respect to that which is
in devel currently.]

Also, since each reference their respective stable baseline repos,
which are still in maintenance and not static content, we use alternate
reference instead of pack reference back to stable.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>