kernel: make linux-yocto-dev recipe use shared source
At the moment, linux-yocto-dev is set up to reference v5.12-rt:

        mainline
           ---> stable-5.12
              ---> linux-rt-5.12
                 ---> linux-yocto-dev

We'll probably move to v5.13-rcN in late May and at that point we
can point it at referencing mainline directly while the respective
stable and -rt versions take shape, or perhaps create placeholders
in advance (as was just done for stable-5.12).

But the good thing is that we'll get v5.13-rcN via mainline, which is
already in the above chain, so we can take our time; there is no
pressure to reconsider the chaining of repos synchronously with the
kernel uprev.

And it is also constructive to consider the reverse case: where we are
pointing at mainline directly and a stable (or -rt) gets added/used.

        mainline
           ---> linux-yocto-dev

In this case, if linux-yocto-dev used a shiny new v5.13.1 stable,
without it yet being explicitly a part of the above chain, then the git
objects relating to that would simply flow from the server directly into
our linux-yocto-dev clone directly.  Nothing breaks or falls over.  So
there is no requirement to immediately react to creation of v5.13 stable
or preempt-rt recipes (even though they are trivial).

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>