bitbake: fetch2/git: append new altref line if/when SRC_URI changed value
A new repository that is cloned with "--reference" in turn creates an
objects/info/alternates file in the new repo containing the path back to
the reference repo objects.  This is normally a "one-and-done" type
operation and the reference never changes.

However we can envision where a package leaves behind a dead end
"stable" fork and re-aligns itself with the main branch, or a newer
stable fork.  In the case of kernel, a real world example would be
leaving behind v5.4.x in favour of v5.10.x for a baseline.

This commit is a consideration for those users - with an existing copy
in their downloads dir that is using a reference that has since been
updated in the SRC_URI of the package recipe.

If we removed the old reference and replaced it, we would break any old
build directories that had branches based on v5.4.x specific commits, so
that is clearly not an option.

If we simply did nothing, things would still work just fine for them;
any v5.10.x objects not found in the v5.4.x reference would be simply
obtained automatically from the server during a fetch update, and placed
in the endpoint repo that is using the older v5.4.x as a reference.

However, since alternates supports a multi-line format, we can simply
append the new reference.  This leaves any v5.4.x build "consumers"
functioning as-is, and also gives a path back to v5.10.x objects in the
new reference, so they need not be duplicated over into the endpoint repo.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>